move utils to the lib folder
