+++ src/components/modals/Open/Open.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useState } from 'react';
 import { observer } from 'mobx-react-lite';
-import { Cookie, loadStory } from '../../../utils';
+import { Cookie, loadStory } from '../../../lib/utils';
 import { Store } from '../../../store';
 import { OpenModalBody } from './OpenBody';
 import { OpenModalActions } from './OpenActions';
