+++ src/components/pages/Inspector/InspectorJSON/InspectorJSON.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useCallback } from 'react';
-import { CodeHighlighter } from '../../../../utils/CodeHighlighter';
+import { CodeHighlighter } from '../../../../lib/utils';
 
 export const InspectorJSON = ({ features }) => {
   const spreadOriginals = useCallback(
