+++ src/components/pages/Splash/Splash.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { observer } from 'mobx-react-lite';
 import { Store } from '../../../store';
 import { DataStoryWidget } from '../../widgets/DataStory';
-import { loadDemo, loadStory } from '../../../utils';
+import { loadDemo, loadStory } from '../../../lib/utils';
 
 interface Props {
   store: Store;
