+++ src/components/widgets/OutputNode/OutputNode.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useState } from 'react';
 import Modal from 'react-modal';
-import { modalStyle } from '../../../utils/modalStyle';
+import { modalStyle } from '../../../lib/utils';
 import { Store, withStore } from '../../../store';
 
 import { OutputNodeWidgetHeader } from './OutputNodeHeader';
