+++ src/diagram/models/DiagramModel.ts
@@ -4,7 +4,7 @@ import {
 } from '@projectstorm/react-diagrams';
 import { NodeModel } from './NodeModel';
 import { SerializedReactDiagram } from '../../lib/types';
-import { version as VERSION } from '../../utils';
+import { version as VERSION } from '../../lib/utils';
 import { stringify } from 'flatted';
 
 /**
