+++ src/lib/utils/version.ts
@@ -1,0 +1,3 @@
+import packageInfo from '../../../package.json';
+
+export const version = packageInfo.version;
