--- src/utils/index.tsx
@@ -1,4 +1,0 @@
-export { renameKey } from './helpers';
-export * from './stories';
-export * from './Cookie';
-export * from './version';
