--- src/utils/version.ts
@@ -1,3 +1,0 @@
-import packageInfo from '../../package.json';
-
-export const version = packageInfo.version;
